pystol-operator/requirements.txt
pyyaml
requests
ansible
openshift
j2cli
jinja2
